Parameter Make.1-String

type t

The type for strings.

val empty : t

The empty string.

val length : t -> int

Returns the length of the string.

val append : t -> t -> t

Concatenates two strings.

val lowercase : t -> t

New string with uppercase letter translated to lowercase (correctness is only needed for ASCII code point).

val iter : (int -> unit) -> t -> unit

Iterates over the unicode code point of the given string.

val of_string : std_string -> t

String from an OCaml string.

val to_utf_8 : ('a -> std_string -> 'a) -> 'a -> t -> 'a

to_utf_8 f v s, is f (... (f (f v s1) s2) ...) sn. Where the concatenation of s1, s2, ... sn is s as an UTF-8 stream.

val compare : t -> t -> int

String comparison. Binary comparison is sufficent.